Update and partially reorganize `valence_client`
