+++ crates/valence/examples/combat.rs
@@ -96,7 +96,7 @@ fn handle_combat_events(
     server: Res<Server>,
     mut clients: Query<CombatQuery>,
     mut sprinting: EventReader<Sprinting>,
-    mut interact_entity: EventReader<InteractEntity>,
+    mut interact_entity: EventReader<InteractEntityEvent>,
 ) {
     for &Sprinting { client, state } in sprinting.iter() {
         if let Ok(mut client) = clients.get_mut(client) {
@@ -104,7 +104,7 @@ fn handle_combat_events(
         }
     }
 
-    for &InteractEntity {
+    for &InteractEntityEvent {
         client: attacker_client,
         entity: victim_client,
         ..
