+++ crates/valence_client/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 anyhow.workspace = true
 bevy_app.workspace = true
 bevy_ecs.workspace = true
+bitfield-struct.workspace = true
 bytes.workspace = true
 glam.workspace = true
 rand.workspace = true
@@ -17,5 +18,5 @@ valence_core.workspace = true
 valence_dimension.workspace = true
 valence_entity.workspace = true
 valence_instance.workspace = true
+valence_nbt.workspace = true
 valence_registry.workspace = true
-
