+++ crates/valence_client/src/keepalive.rs
@@ -1,4 +1,4 @@
-use valence_core::packet::c2s::play::KeepAliveC2s;
+use valence_core::protocol::{packet_id, Decode, Encode, Packet};
 use valence_core::CoreSettings;
 
 use super::*;
@@ -80,4 +80,16 @@ fn handle_keepalive_response(
             }
         }
     }
+}
+
+#[derive(Copy, Clone, Debug, Encode, Decode, Packet)]
+#[packet(id = packet_id::KEEP_ALIVE_C2S)]
+pub struct KeepAliveC2s {
+    pub id: u64,
+}
+
+#[derive(Copy, Clone, Debug, Encode, Decode, Packet)]
+#[packet(id = packet_id::KEEP_ALIVE_S2C)]
+pub struct KeepAliveS2c {
+    pub id: u64,
 }
