+++ crates/valence_client/src/lib.rs
@@ -29,6 +29,11 @@ use bevy_ecs::query::WorldQuery;
 use bevy_ecs::system::Command;
 use bytes::{Bytes, BytesMut};
 use glam::{DVec3, Vec3};
+use packet::{
+    DeathMessageS2c, DisconnectS2c, GameEventKind, GameJoinS2c, GameStateChangeS2c,
+    OverlayMessageS2c, PlayerRespawnS2c, PlayerSpawnPositionS2c, PlayerSpawnS2c, SubtitleS2c,
+    TitleFadeS2c, TitleS2c,
+};
 use rand::Rng;
 use tracing::warn;
 use uuid::Uuid;
@@ -38,45 +43,49 @@ use valence_core::chunk_pos::{ChunkPos, ChunkView};
 use valence_core::despawn::Despawned;
 use valence_core::game_mode::GameMode;
 use valence_core::ident::Ident;
-use valence_core::packet::byte_angle::ByteAngle;
-use valence_core::packet::encode::{PacketEncoder, WritePacket};
-use valence_core::packet::global_pos::GlobalPos;
-use valence_core::packet::s2c::play::game_state_change::GameEventKind;
-use valence_core::packet::s2c::play::particle::Particle;
-use valence_core::packet::s2c::play::player_position_look::Flags as PlayerPositionLookFlags;
-use valence_core::packet::s2c::play::{
-    ChunkLoadDistanceS2c, ChunkRenderDistanceCenterS2c, CustomPayloadS2c, DeathMessageS2c,
-    DisconnectS2c, EntitiesDestroyS2c, EntitySetHeadYawS2c, EntitySpawnS2c, EntityStatusS2c,
-    EntityTrackerUpdateS2c, EntityVelocityUpdateS2c, ExperienceOrbSpawnS2c, GameJoinS2c,
-    GameMessageS2c, GameStateChangeS2c, KeepAliveS2c, OverlayMessageS2c, ParticleS2c, PlaySoundS2c,
-    PlayerActionResponseS2c, PlayerPositionLookS2c, PlayerRespawnS2c, PlayerSpawnPositionS2c,
-    PlayerSpawnS2c, ResourcePackSendS2c, SubtitleS2c, TitleFadeS2c, TitleS2c, UnloadChunkS2c,
-};
-use valence_core::packet::var_int::VarInt;
-use valence_core::packet::Packet;
+use valence_core::particle::{Particle, ParticleS2c};
 use valence_core::property::Property;
+use valence_core::protocol::byte_angle::ByteAngle;
+use valence_core::protocol::encode::{PacketEncoder, WritePacket};
+use valence_core::protocol::global_pos::GlobalPos;
+use valence_core::protocol::packet::sound::{PlaySoundS2c, Sound, SoundCategory};
+use valence_core::protocol::var_int::VarInt;
+use valence_core::protocol::Packet;
 use valence_core::scratch::ScratchBuf;
-use valence_core::sound::{Sound, SoundCategory};
 use valence_core::text::Text;
 use valence_core::uuid::UniqueId;
 use valence_core::Server;
+use valence_entity::packet::{
+    EntitiesDestroyS2c, EntitySetHeadYawS2c, EntitySpawnS2c, EntityStatusS2c,
+    EntityTrackerUpdateS2c, EntityVelocityUpdateS2c, ExperienceOrbSpawnS2c,
+};
 use valence_entity::player::PlayerEntityBundle;
 use valence_entity::{
     ClearEntityChangesSet, EntityId, EntityKind, EntityStatus, HeadYaw, Location, Look, ObjectData,
     OldLocation, OldPosition, OnGround, PacketByteRange, Position, TrackedData, Velocity,
 };
+use valence_instance::packet::{
+    ChunkLoadDistanceS2c, ChunkRenderDistanceCenterS2c, UnloadChunkS2c,
+};
 use valence_instance::{ClearInstanceChangesSet, Instance, WriteUpdatePacketsToInstancesSet};
 use valence_registry::{RegistryCodec, RegistryCodecSet};
 
 pub mod action;
+pub mod chat;
 pub mod command;
+pub mod custom_payload;
 pub mod event_loop;
+pub mod hand_swing;
+pub mod interact_block;
 pub mod interact_entity;
+pub mod interact_item;
 pub mod keepalive;
-pub mod misc;
 pub mod movement;
+pub mod op_level;
 pub mod packet;
+pub mod resource_pack;
 pub mod settings;
+pub mod status;
 pub mod teleport;
 pub mod weather;
 
@@ -116,7 +125,6 @@ impl Plugin for ClientPlugin {
                 update_game_mode,
                 update_tracked_data.after(WriteUpdatePacketsToInstancesSet),
                 init_tracked_data.after(WriteUpdatePacketsToInstancesSet),
-                update_op_level,
             )
                 .in_set(UpdateClientsSet),
         )
@@ -137,10 +145,17 @@ impl Plugin for ClientPlugin {
         keepalive::build(app);
         interact_entity::build(app);
         settings::build(app);
-        misc::build(app);
         action::build(app);
         teleport::build(app);
         weather::build(app);
+        chat::build(app);
+        custom_payload::build(app);
+        hand_swing::build(app);
+        interact_block::build(app);
+        interact_item::build(app);
+        op_level::build(app);
+        resource_pack::build(app);
+        status::build(app);
     }
 }
 
@@ -157,7 +172,7 @@ pub struct ClientBundle {
     pub properties: Properties,
     pub compass_pos: CompassPos,
     pub game_mode: GameMode,
-    pub op_level: OpLevel,
+    pub op_level: op_level::OpLevel,
     pub action_sequence: action::ActionSequence,
     pub view_distance: ViewDistance,
     pub old_view_distance: OldViewDistance,
@@ -190,7 +205,7 @@ impl ClientBundle {
             properties: Properties(args.properties),
             compass_pos: CompassPos::default(),
             game_mode: GameMode::default(),
-            op_level: OpLevel::default(),
+            op_level: op_level::OpLevel::default(),
             action_sequence: action::ActionSequence::default(),
             view_distance: ViewDistance::default(),
             old_view_distance: OldViewDistance(2),
@@ -308,22 +323,6 @@ impl Client {
         }
     }
 
-    /// Sends a system message to the player which is visible in the chat. The
-    /// message is only visible to this client.
-    pub fn send_message(&mut self, msg: impl Into<Text>) {
-        self.write_packet(&GameMessageS2c {
-            chat: msg.into().into(),
-            overlay: false,
-        });
-    }
-
-    pub fn send_custom_payload(&mut self, channel: Ident<&str>, data: &[u8]) {
-        self.write_packet(&CustomPayloadS2c {
-            channel: channel.into(),
-            data: data.into(),
-        });
-    }
-
     /// Kills the client and shows `message` on the death screen. If an entity
     /// killed the player, you should supply it as `killer`.
     pub fn kill(&mut self, killer: Option<EntityId>, message: impl Into<Text>) {
@@ -339,31 +338,6 @@ impl Client {
         self.write_packet(&GameStateChangeS2c {
             kind: GameEventKind::WinGame,
             value: if show_credits { 1.0 } else { 0.0 },
-        });
-    }
-
-    /// Requests that the client download and enable a resource pack.
-    ///
-    /// # Arguments
-    /// * `url` - The URL of the resource pack file.
-    /// * `hash` - The SHA-1 hash of the resource pack file. Any value other
-    ///   than a 40-character hexadecimal string is ignored by the client.
-    /// * `forced` - Whether a client should be kicked from the server upon
-    ///   declining the pack (this is enforced client-side)
-    /// * `prompt_message` - A message to be displayed with the resource pack
-    ///   dialog.
-    pub fn set_resource_pack(
-        &mut self,
-        url: &str,
-        hash: &str,
-        forced: bool,
-        prompt_message: Option<Text>,
-    ) {
-        self.write_packet(&ResourcePackSendS2c {
-            url,
-            hash,
-            forced,
-            prompt_message: prompt_message.map(|t| t.into()),
         });
     }
 
@@ -593,20 +567,6 @@ pub struct Ip(pub IpAddr);
 /// The position that regular compass items will point to.
 #[derive(Component, Copy, Clone, PartialEq, Eq, Default, Debug)]
 pub struct CompassPos(pub BlockPos);
-
-#[derive(Component, Clone, PartialEq, Eq, Default, Debug)]
-pub struct OpLevel(u8);
-
-impl OpLevel {
-    pub fn get(&self) -> u8 {
-        self.0
-    }
-
-    /// Sets the op level. Value is clamped to `0..=3`.
-    pub fn set(&mut self, lvl: u8) {
-        self.0 = lvl.min(3);
-    }
-}
 
 #[derive(Component, Clone, PartialEq, Eq, Debug)]
 pub struct ViewDistance(u8);
@@ -1268,14 +1228,5 @@ fn update_tracked_data(mut clients: Query<(&mut Client, &TrackedData)>) {
                 metadata: update_data.into(),
             });
         }
-    }
-}
-
-fn update_op_level(mut clients: Query<(&mut Client, &OpLevel), Changed<OpLevel>>) {
-    for (mut client, lvl) in &mut clients {
-        client.write_packet(&EntityStatusS2c {
-            entity_id: 0,
-            entity_status: 24 + lvl.0,
-        });
     }
 }
