+++ crates/valence_client/src/teleport.rs
@@ -1,4 +1,4 @@
-use valence_core::packet::c2s::play::TeleportConfirmC2s;
+use packet::*;
 
 use super::*;
 use crate::event_loop::{EventLoopSchedule, EventLoopSet, PacketEvent};
@@ -79,7 +79,7 @@ fn teleport(
                 yaw: if changed_yaw { look.yaw } else { 0.0 },
                 pitch: if changed_pitch { look.pitch } else { 0.0 },
                 flags,
-                teleport_id: VarInt(state.teleport_id_counter as i32),
+                teleport_id: (state.teleport_id_counter as i32).into(),
             });
 
             state.pending_teleports = state.pending_teleports.wrapping_add(1);
@@ -120,5 +120,40 @@ fn handle_teleport_confirmations(
                 }
             }
         }
+    }
+}
+
+pub mod packet {
+    use bitfield_struct::bitfield;
+    use glam::DVec3;
+    use valence_core::protocol::var_int::VarInt;
+    use valence_core::protocol::{packet_id, Decode, Encode, Packet};
+
+    #[derive(Copy, Clone, Debug, Encode, Decode, Packet)]
+    #[packet(id = packet_id::TELEPORT_CONFIRM_C2S)]
+    pub struct TeleportConfirmC2s {
+        pub teleport_id: VarInt,
+    }
+
+    #[derive(Copy, Clone, PartialEq, Debug, Encode, Decode, Packet)]
+    #[packet(id = packet_id::PLAYER_POSITION_LOOK_S2C)]
+    pub struct PlayerPositionLookS2c {
+        pub position: DVec3,
+        pub yaw: f32,
+        pub pitch: f32,
+        pub flags: PlayerPositionLookFlags,
+        pub teleport_id: VarInt,
+    }
+
+    #[bitfield(u8)]
+    #[derive(PartialEq, Eq, Encode, Decode)]
+    pub struct PlayerPositionLookFlags {
+        pub x: bool,
+        pub y: bool,
+        pub z: bool,
+        pub y_rot: bool,
+        pub x_rot: bool,
+        #[bits(3)]
+        _pad: u8,
     }
 }
